+++ tests/ReadWriteContextTest.php
@@ -101,6 +101,7 @@ class ReadWriteContextTest extends TestCase
                 'unsetThroughAlias' => ['', "wrote to \$GLOBALS['removed']"],
                 'rebindAlias' => ['', "wrote to \$GLOBALS['second']"],
                 'dynamicGlobalKey' => ['', 'wrote to $GLOBALS[$key]'],
+                'parameterReference' => ['', 'wrote to argument $value'],
                 'nonGlobalsReference' => ['read from superglobal $_SESSION', ''],
             ],
             $this->reportedRowsAtPath(self::REFERENCE_ALIAS_FIXTURE, ['--strict', '--props']),